[5.0.32][linux-capi] Return int pointers as jlong, to support 64bit address space WARNING: The native methode interface probably needs to be changed to long as well! SEE ALSO: http://stackoverflow.com/questions/337268/what-is-the-correct-way-to-store-a-native-pointer-inside-a-java-object

+++ modules/capi/c++/linux/de_powerisdnmonitor_capi_PIMCapi.c
@@ -183,7 +183,7 @@ JNIEXPORT jint JNICALL Java_de_powerisdnmonitor_capi_PIMCapi_nGetAddress
 	jbyte *jap;
 	jap = env->GetByteArrayElements(ja, 0);
 	//(*env)->ReleaseByteArrayElements(env, ja, jap, 0);
-	return (jint)jap;
+	return (jlong)jap;
 }  
 
 JNIEXPORT jbyteArray JNICALL Java_de_powerisdnmonitor_capi_PIMCapi_nGetData
